+++ src/Cuemon.Threading/TaskFuncFactory.cs
@@ -17,7 +17,7 @@ namespace Cuemon.Threading
         /// <returns>An instance of <see cref="TaskFuncFactory{TTuple,TResult}"/> object initialized with the specified <paramref name="method"/>.</returns>
         public static TaskFuncFactory<Template, TResult> Create<TResult>(Func<CancellationToken, Task<TResult>> method)
         {
-            return new TaskFuncFactory<Template, TResult>((tuple, token) => method(token), Template.CreateZero(), method);
+            return new TaskFuncFactory<Template, TResult>((_, token) => method(token), Template.CreateZero(), method);
         }
 
         /// <summary>
@@ -423,69 +423,6 @@ namespace Cuemon.Threading
         public static TaskFuncFactory<Template<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15>, TResult> Create<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, TResult>(Func<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, CancellationToken, Task<TResult>> method, T1 arg1, T2 arg2, T3 arg3, T4 arg4, T5 arg5, T6 arg6, T7 arg7, T8 arg8, T9 arg9, T10 arg10, T11 arg11, T12 arg12, T13 arg13, T14 arg14, T15 arg15)
         {
             return new TaskFuncFactory<Template<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15>, TResult>((tuple, token) => method(tuple.Arg1, tuple.Arg2, tuple.Arg3, tuple.Arg4, tuple.Arg5, tuple.Arg6, tuple.Arg7, tuple.Arg8, tuple.Arg9, tuple.Arg10, tuple.Arg11, tuple.Arg12, tuple.Arg13, tuple.Arg14, tuple.Arg15, token), Template.CreateFifteen(arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10, arg11, arg12, arg13, arg14, arg15), method);
-        }
-    }
-
-    /// <summary>
-    /// Provides an easy way of invoking an <see cref="Func{TResult}" /> function delegate regardless of the amount of parameters provided.
-    /// </summary>
-    /// <typeparam name="TTuple">The type of the n-tuple representation of a <see cref="Template"/>.</typeparam>
-    /// <typeparam name="TResult">The type of the return value of the function delegate <see cref="Method"/>.</typeparam>
-    public sealed class TaskFuncFactory<TTuple, TResult> : TemplateFactory<TTuple> where TTuple : Template
-    {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="TaskFuncFactory{TTuple,TResult}"/> class.
-        /// </summary>
-        /// <param name="method">The function delegate to invoke.</param>
-        /// <param name="tuple">The n-tuple argument of <paramref name="method"/>.</param>
-        public TaskFuncFactory(Func<TTuple, CancellationToken, Task<TResult>> method, TTuple tuple) : this(method, tuple, method)
-        {
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="TaskFuncFactory{TTuple,TResult}"/> class.
-        /// </summary>
-        /// <param name="method">The function delegate to invoke.</param>
-        /// <param name="tuple">The n-tuple argument of <paramref name="method"/>.</param>
-        /// <param name="originalDelegate">The original delegate wrapped by <paramref name="method"/>.</param>
-        internal TaskFuncFactory(Func<TTuple, CancellationToken, Task<TResult>> method, TTuple tuple, Delegate originalDelegate) : base(tuple, originalDelegate != null)
-        {
-            Method = method;
-            DelegateInfo = Decorator.RawEnclose(method).ResolveDelegateInfo(originalDelegate);
-        }
-
-        /// <summary>
-        /// Gets the function delegate to invoke.
-        /// </summary>
-        /// <value>The function delegate to invoke.</value>
-        private Func<TTuple, CancellationToken, Task<TResult>> Method { get; set; }
-
-        /// <summary>
-        /// Executes the function delegate associated with this instance.
-        /// </summary>
-        /// <param name="ct">The token to monitor for cancellation requests. The default value is <see cref="CancellationToken.None"/>.</param>
-        /// <returns>A task that represents the asynchronous operation. The task result contains the return value of the function delegate associated with this instance.</returns>
-        /// <exception cref="InvalidOperationException">
-        /// No delegate was specified on the factory.
-        /// </exception>
-        /// <exception cref="OperationCanceledException">
-        /// The <paramref name="ct"/> was canceled.
-        /// </exception>
-        public Task<TResult> ExecuteMethodAsync(CancellationToken ct)
-        {
-            ThrowIfNoValidDelegate(Condition.IsNull(Method));
-            ct.ThrowIfCancellationRequested();
-            return Method.Invoke(GenericArguments, ct);
-        }
-
-        /// <summary>
-        /// Creates a shallow copy of the current <see cref="TaskFuncFactory{TTuple,TResult}"/> object.
-        /// </summary>
-        /// <returns>A new <see cref="TaskFuncFactory{TTuple,TResult}"/> that is a copy of this instance.</returns>
-        /// <remarks>When thread safety is required this is the method to invoke.</remarks>
-        public override TemplateFactory<TTuple> Clone()
-        {
-            return new TaskFuncFactory<TTuple, TResult>(Method, GenericArguments.Clone() as TTuple);
         }
     }
 }
